+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+##
+- Revert remove `DCC_RX_MIN_CV_WRITE_PACKETS` and add `DCC_STANDARD_COMPLIANCE` CMake option
+
 ## 0.43.0
 - Remove `DCC_STANDARD_COMPLIANCE` CMake option
 - Add `DCC_RX_MIN_CV_WRITE_PACKETS` CMake option
@@ -14,7 +17,7 @@
 - Add `bidi::Dissector` ([#59](https://github.com/ZIMO-Elektronik/DCC/issues/59))
 - Add `DCC_STANDARD_COMPLIANCE` CMake option ([#78](https://github.com/ZIMO-Elektronik/DCC/issues/78))
 - Bugfix receiver sends channel 1 replies to broadcast packets ([#60](https://github.com/ZIMO-Elektronik/DCC/issues/60))
-- Bugfix CV29:1 did not work with consist ([#68](https://github.com/ZIMO-Elektronik/DCC/issues/68))
+- Bugfix CV29:0 did not work with consist ([#68](https://github.com/ZIMO-Elektronik/DCC/issues/68))
 - Bugfix CV short access for CV23/24 ([#69](https://github.com/ZIMO-Elektronik/DCC/issues/69))
 - Bugfix idle packet iterator not updated ([#71](https://github.com/ZIMO-Elektronik/DCC/issues/71))
 - Bugfix accessory address encoding ([#74](https://github.com/ZIMO-Elektronik/DCC/issues/74))